Image resizer transformer now encapsulates all the necessary data and functionality for cropping and resizing.
